Did small refector, all util function for inference api call are now in the same file
