+++ src/utils/__init__.py
@@ -5,6 +5,4 @@ from .model_registry import get_model
 from .path import (get_best_checkpoint, get_run_dir, read_config, resolve_path,
                    write_config)
 from .timing import Timer
-from .load_model import load_model
-from .make_prediction import make_prediction
-from .handle_deployment import handle_input_inference, handle_output_inference
+from .handle_deployment import handle_input_inference, handle_output_inference, load_model, make_prediction
